Update ytdown.py +mp3 +fixBug +startFromTheMid
I am using juvekaradheesh/youtube-playlist-downloader code, to add download as mp3 file. I fixed a bug when a video appears more than once in the playlist (the problem was when the rename() function was activated, so I just added the number of the video to its name). I also added an option to start the download from a specific video (when the code crashes and the user doesn't want to start downloading everything from the beginning - the videos are numbered for the sake of the process).